Fix the ocult interpreter to actually implement LHMN
It probably makes more sense to document the difference instead of
fixing it, in an updated version, since changing the behavior could
break somebody's solutions to the aspects puzzles.

It didn't affect my solutions, since I just made sure that the rules
only every applied once, and I suspect that *most* people did that,
but still.

I think that I might be the only person to have ever done the
2d/aspects puzzle, given that it crashed when I solved it, so that
side of it is less scary.